+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This can also be enabled programmatically with `warnings.simplefilter('default',
 - some documentation on how to embed Matplotlib charts: [Maths](https://pyfpdf.github.io/fpdf2/Maths.html)
 ### Fixed
 - infinite loop when calling `.multi_cell()` without enough horizontal space - _cf._ [#389](https://github.com/PyFPDF/fpdf2/issues/389)
+### Removed
+- support for `.pkl` passed to `add_font()`. This was deprecated since v2.5.1.
 
 ## [2.5.2] - 2022-04-13
 ### Added